api/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
cassandra-driver==3.29.1
certifi==2024.8.30
click==8.1.7
coverage==7.6.8
fastapi==0.115.0
flake8==7.1.1
geomet==0.2.1.post1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==6.0.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
